update アセット 検索

+++ src/main/java/co/assets/manage/dto/req/QueryAssetRequest.java
@@ -18,7 +18,7 @@ public record QueryAssetRequest(
         Integer pageSize,
 
         //上页最大ID
-        Integer lastPageMaxId
+        Long lastPageMaxId
 ) {
 
     public QueryAssetRequest {
